pyproject.toml
[project]
name = "aic2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.129.0",
    "google-genai>=1.63.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
    "xai-sdk>=1.7.0",
]
